+++ src/dk/aau/cs/ds306e18/tournament/ui/BracketOverviewTabController.java
@@ -1,6 +1,7 @@
 package dk.aau.cs.ds306e18.tournament.ui;
 
 import dk.aau.cs.ds306e18.tournament.model.format.StageStatusChangeListener;
+import dk.aau.cs.ds306e18.tournament.model.match.MatchResultDependencyException;
 import dk.aau.cs.ds306e18.tournament.model.match.Series;
 import dk.aau.cs.ds306e18.tournament.rlbot.MatchRunner;
 import dk.aau.cs.ds306e18.tournament.model.Bot;
@@ -41,13 +42,14 @@ public class BracketOverviewTabController implements StageStatusChangeListener, 
     @FXML private Button modifyConfigBtn;
     @FXML private Button editMatchBtn;
     @FXML private Button switchColorsBtn;
+    @FXML private Button extendSeriesBtn;
+    @FXML private Button shortenSeriesBtn;
     @FXML private Label blueTeamNameLabel;
     @FXML private ListView<Bot> blueTeamListView;
     @FXML private Label orangeTeamNameLabel;
     @FXML private ListView<Bot> orangeTeamListView;
     @FXML private ScrollPane overviewScrollPane;
     @FXML private GridPane selectedMatchInfo;
-    @FXML private HBox selectedMatchButtonHolder;
     @FXML private HBox stageNavigationButtonsHolder;
     @FXML private Button nextStageBtn;
     @FXML private Button prevStageBtn;
@@ -62,7 +64,7 @@ public class BracketOverviewTabController implements StageStatusChangeListener, 
     private int showedStageIndex = -1;
     private ModelCoupledUI coupledBracket;
     private Format showedFormat;
-    private SeriesVisualController selectedMatch;
+    private SeriesVisualController selectedSeries;
 
     @FXML
     private void initialize() {
@@ -124,7 +126,7 @@ public class BracketOverviewTabController implements StageStatusChangeListener, 
             showedStageIndex = -1;
             showedFormat = null;
             showStartTournamentInstructions(true);
-            setSelectedMatch(null);
+            setSelectedSeries(null);
             updateStageNavigationButtons();
         } else {
             if (showedStageIndex == -1) showedStageIndex = 0;
@@ -133,7 +135,7 @@ public class BracketOverviewTabController implements StageStatusChangeListener, 
             updateStageNavigationButtons();
         }
 
-        updateTeamViewer(selectedMatch == null ? null : selectedMatch.getShowedSeries());
+        updateTeamViewer(selectedSeries == null ? null : selectedSeries.getShowedSeries());
     }
 
     /** @return a string that contains text describing the requirements for starting the tournament. */
@@ -217,14 +219,14 @@ public class BracketOverviewTabController implements StageStatusChangeListener, 
     /**
      * Sets the selected match.
      */
-    public void setSelectedMatch(SeriesVisualController match) {
-        if (selectedMatch != null){
-            selectedMatch.getShowedSeries().unregisterMatchChangeListener(this);
-            selectedMatch.getRoot().getStyleClass().remove("selectedMatch");
+    public void setSelectedSeries(SeriesVisualController match) {
+        if (selectedSeries != null){
+            selectedSeries.getShowedSeries().unregisterMatchChangeListener(this);
+            selectedSeries.getRoot().getStyleClass().remove("selectedMatch");
         }
-        this.selectedMatch = match;
+        this.selectedSeries = match;
         updateTeamViewer(match == null ? null : match.getShowedSeries());
-        if(selectedMatch != null) { selectedMatch.getShowedSeries().registerMatchChangeListener(this); }
+        if(selectedSeries != null) { selectedSeries.getShowedSeries().registerMatchChangeListener(this); }
     }
 
     /**
@@ -233,7 +235,7 @@ public class BracketOverviewTabController implements StageStatusChangeListener, 
     @FXML
     void deselectOnRightClick(MouseEvent event) {
         if (event.getButton().equals(MouseButton.SECONDARY)){
-            setSelectedMatch(null);
+            setSelectedSeries(null);
         }
     }
 
@@ -244,7 +246,6 @@ public class BracketOverviewTabController implements StageStatusChangeListener, 
         boolean disable = (series == null);
         botInfoBox.setVisible(false);
         selectedMatchInfo.setDisable(disable);
-        selectedMatchButtonHolder.setDisable(disable);
 
         if (series != null && series.getBlueTeam() != null) {
             // Blue team
@@ -274,7 +275,7 @@ public class BracketOverviewTabController implements StageStatusChangeListener, 
 
     /** Disables/Enables the play and edit match buttons */
     public void updateMatchPlayAndEditButtons() {
-        if (selectedMatch == null || selectedMatch.getShowedSeries().getStatus() == Series.Status.NOT_PLAYABLE) {
+        if (selectedSeries == null || selectedSeries.getShowedSeries().getStatus() == Series.Status.NOT_PLAYABLE) {
             editMatchBtn.setDisable(true);
             playMatchBtn.setDisable(true);
             modifyConfigBtn.setDisable(true);
@@ -285,7 +286,15 @@ public class BracketOverviewTabController implements StageStatusChangeListener, 
             playMatchBtn.setDisable(false);
             modifyConfigBtn.setDisable(false);
             // color switching is disabled when match has been played to avoid confusion when comparing replays and bracket
-            switchColorsBtn.setDisable(selectedMatch.getShowedSeries().hasBeenPlayed());
+            switchColorsBtn.setDisable(selectedSeries.getShowedSeries().hasBeenPlayed());
+        }
+
+        if (selectedSeries == null) {
+            extendSeriesBtn.setDisable(true);
+            shortenSeriesBtn.setDisable(true);
+        } else {
+            extendSeriesBtn.setDisable(false);
+            shortenSeriesBtn.setDisable(selectedSeries.getShowedSeries().getSeriesLength() == 1);
         }
     }
 
@@ -309,7 +318,7 @@ public class BracketOverviewTabController implements StageStatusChangeListener, 
             FXMLLoader loader = new FXMLLoader(BracketOverviewTabController.class.getResource("layout/EditSeriesScore.fxml"));
             AnchorPane editMatchStageRoot = loader.load();
             EditSeriesScoreController emsc = loader.getController();
-            emsc.setSeries(selectedMatch.getShowedSeries());
+            emsc.setSeries(selectedSeries.getShowedSeries());
             editMatchScoreStage.setScene(new Scene(editMatchStageRoot));
 
             // Calculate the center position of the main window.
@@ -362,13 +371,13 @@ public class BracketOverviewTabController implements StageStatusChangeListener, 
 
     @Override
     public void onMatchChanged(Series series) {
-        updateTeamViewer(selectedMatch.getShowedSeries());
+        updateTeamViewer(selectedSeries.getShowedSeries());
     }
 
     public void prevStageBtnOnAction(ActionEvent actionEvent) {
         showedStageIndex--;
         showStage(Tournament.get().getStages().get(showedStageIndex));
-        setSelectedMatch(null);
+        setSelectedSeries(null);
         updateStageNavigationButtons();
     }
 
@@ -381,16 +390,16 @@ public class BracketOverviewTabController implements StageStatusChangeListener, 
         showedStageIndex++;
         showStage(Tournament.get().getStages().get(showedStageIndex));
 
-        setSelectedMatch(null);
+        setSelectedSeries(null);
         updateStageNavigationButtons();
     }
 
     public void onPlayMatchBtnAction(ActionEvent actionEvent) {
-        MatchRunner.startMatch(Tournament.get().getRlBotSettings().getMatchConfig(), selectedMatch.getShowedSeries());
+        MatchRunner.startMatch(Tournament.get().getRlBotSettings().getMatchConfig(), selectedSeries.getShowedSeries());
     }
 
     public void modifyConfigButtonOnAction(ActionEvent actionEvent) {
-        boolean ready = MatchRunner.prepareMatch(Tournament.get().getRlBotSettings().getMatchConfig(), selectedMatch.getShowedSeries());
+        boolean ready = MatchRunner.prepareMatch(Tournament.get().getRlBotSettings().getMatchConfig(), selectedSeries.getShowedSeries());
         if (ready) {
             Alerts.infoNotification("Modified config file", "The rlbot.cfg was successfully modified to the selected match.");
         }
@@ -406,6 +415,36 @@ public class BracketOverviewTabController implements StageStatusChangeListener, 
     }
 
     public void onSwitchColorsBtnAction(ActionEvent actionEvent) {
-        selectedMatch.getShowedSeries().setTeamOneToBlue(!selectedMatch.getShowedSeries().isTeamOneBlue());
+        selectedSeries.getShowedSeries().setTeamOneToBlue(!selectedSeries.getShowedSeries().isTeamOneBlue());
+    }
+
+    public void extendSeriesBtnOnAction(ActionEvent actionEvent) {
+        carefullyChangeSeriesLength(selectedSeries.getShowedSeries().getSeriesLength() + 2);
+    }
+
+    public void shortenSeriesBtnOnAction(ActionEvent actionEvent) {
+        // We assume the button is only clickable if the match length is greater than 1
+        carefullyChangeSeriesLength(selectedSeries.getShowedSeries().getSeriesLength() - 2);
+    }
+
+    /**
+     * Changes the selected series length. In case this change changes the outcome of the series (and other series
+     * depends on this outcome) an alert prompt is shown, and the user must confirm if they want to proceed.
+     */
+    public void carefullyChangeSeriesLength(int length) {
+        Series series = selectedSeries.getShowedSeries();
+        boolean force = false;
+        boolean cancelled = false;
+        while (!cancelled) {
+            try {
+                series.setSeriesLength(length, force);
+                break;
+            } catch (MatchResultDependencyException e) {
+                // An MatchResultDependencyException is thrown if the outcome has changed and subsequent matches depends on this outcome
+                // Ask if the user wants to proceed
+                force = Alerts.confirmAlert("The outcome of this match has changed", "This change will reset the subsequent matches. Do you want to proceed?");
+                cancelled = !force;
+            }
+        }
     }
 }